+++ android/app/src/main/java/com/wallkraft/app/presentation/common/WallpaperListViewModel.kt
@@ -98,6 +98,7 @@ abstract class WallpaperListViewModel(
                         categories = initialCategories,
                         purity = effectivePurity,
                         sorting = settings.sorting,
+                        topRange = settings.topRange,
                         orientation = settings.orientation,
                         query = initialQuery,
                     ),